Fix 2GiB chunk-store open cliff

+++ src/chunk_store.h
@@ -191,7 +191,7 @@ struct ChunkStore {
 
   int nChunks;
   i64 iIndexOffset;
-  int nIndexSize;
+  i64 nIndexSize;
   i64 iWalOffset;
   i64 iFileSize;
